kiosk image

+++ src/features/CPU-firebase-interaction.js
@@ -38,7 +38,7 @@ const CPUFirebase = {
   },
   async changeKioskImage(file) {
     if (
-      this.userDocData.kiosk_image !== "" ||
+      this.userDocData.kiosk_image !== "" &&
       this.userDocData.kiosk_image !== "defaultBooth/광어.jpg"
     )
       await deleteObject(ref(storage, this.userDocData.kiosk_image));
